requirements.txt
pytest~=7.4.3
pytest-playwright
playwright==1.40.0
black
allure-pytest
pytest-rerunfailures